+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Dodo x.x.x (not yet released)
 * Update to wikimedia/zest-css 2.0.0.
 * Implement the following non-standard methods for PHP compatibility:
-  * Element::setIdAttribute() (stubbed out)
+  * Document::loadXML()
   * Document::saveXML()
+  * DocumentFragment::appendXML()
+  * Element::setIdAttribute() (stubbed out)
 
 # Dodo 0.2.0 (2021-07-26)
 * Update to IDLeDOM 0.7.2.